+++ src/torchfm/dataset/wrapper_multivalued_dataset.py
@@ -20,7 +20,7 @@ class WrapperMultivaluedDataset(torch.utils.data.Dataset):
     def __init__(self, dataset_path, sep=',', secondary_sep='|', engine='c', header='infer', spark=None):
         miltival_col_name = 'genres'
         data = read_pd_dataframe(dataset_path, sep, engine, header)
-        self.targets = data['label'].to_numpy()
+        self.targets = data['label'].to_numpy().astype(np.float32)
         data = data.drop('label', axis=1)
 
         data[miltival_col_name] = data[miltival_col_name].str.split(secondary_sep).apply(lambda x: [int(i) for i in x])
@@ -31,14 +31,14 @@ class WrapperMultivaluedDataset(torch.utils.data.Dataset):
         global_offsets = list(range(num_columns - 1))
 
         self.weights = np.vstack(data[miltival_col_name].apply(lambda x: WrapperMultivaluedDataset._create_weights(x, self.max_length, num_columns - 1)).to_numpy())
-        self.offsets = np.vstack(data[miltival_col_name].apply(lambda x: WrapperMultivaluedDataset._create_offsets(x, num_columns - 2, global_offsets)).to_numpy())
+        self.offsets = np.vstack(data[miltival_col_name].apply(lambda x: WrapperMultivaluedDataset._create_offsets(x, num_columns - 2, global_offsets)).to_numpy().astype(np.int32))
 
         data[miltival_col_name] = data[miltival_col_name].apply(lambda x: x + [0] * (self.max_length - len(x)))
 
         lst_genres = ['genre' + str(i) for i in range(0, 6)]
         df_genres = pd.DataFrame(data[miltival_col_name].to_list(), columns=lst_genres)
 
-        self.indices = pd.concat([data.drop([miltival_col_name], axis=1), df_genres], axis=1, join='inner').to_numpy()
+        self.indices = pd.concat([data.drop([miltival_col_name], axis=1), df_genres], axis=1, join='inner').to_numpy().astype(np.int32)
         self.field_dims = np.max(self.indices)
         self.num_columns = num_columns
